Remove getting snapshots of recording between two different times
This is not currently being used and it is unclear how this can be
used in the future considering that we do not have a good way of
knowing if data has already been overwritten before scheduling this
request.

We may revive this functionality in the future if we feel there is a
need for this.