Introducing Monitor API to track sending pipeline
The end goal is to capture & send health & volume metrics.

Monitor has callbacks for...
- publishing to disruptor
- serializing
- sending to agent

Success and failure for each step are split into separate methods.

One design goal is to have a no cost NoopMonitor be used by default to start to avoid any accidental overhead being introduced.

To that end, I've chosen to pass the DDAgentWriter to the Monitor.  A health Monitor may chose to call additional methods on DDAgentWriter to gather additional information, but the NoopMonitor will not occur any overhead.