Introducing StatsdMonitor
Introducing Statsd variant of Monitor which uses dogstatsd to send metrics

This change leaves the "monitor" hard-coded to a NoopMonitor, but I've confirmed that all tests pass with StatsdMonitor as well.

This change also introduces a direct dependency on dogstatsd, but I'm assuming that won't be kept in the final pull request.