Finishing statsd monitor configuration
Added a new constructor to DDAgentWriter that takes a DDAgentWriter.Monitor.

Added a createMonitor routine to Writer -- used in createAgentWriter.

Made NoopMonitor & StatsdMonitor public for use in Writer (I could also do package visible)

Kept the old DDAgentWriter constructors for use in tests.  Those may or may not be removed before the PR is merged.