Tweaking DDAgentWriter.toString
Decided to not print out monitor information when the monitor is noop.  This fixes some DDAgentWriter tests that I broke in a prior commit.

I think this approach strikes a nice balance of preserving the same output by default, but still giving a clear indication in the logs when health monitoring is enabled.

I decided to check for the NoopMonitor by doing instanceof check.  I don't particularly like this solution, but I think it better than adding an isNoOp routine onto both Monitors or relying on NoopMonitor.toString to return a specifc String (probably empty String).