Adding memory pool Listener
Added an additional AgentBuilder.Listener that updates memory pool usage on class load.

Arguably, we could use the data from JMX fetch, but I don't want to rely on JMX fetch and would like some extra detail.

In the future, we might want to just move this to separate thread or timer, but I think this works well enough for now.