PROF-581: Add exception sampling event
In order to reduce the number of exception events which can become quite overwhelming a statistical sampling is employed to pick semi-random exceptions and generate events only for those. The sampling rate is adaptive and ensures that the number of events in an average recording will converge to a finite, not very large number.