Experimental instrumentation for the OpenTelemetry Tracing API
This can be used as an alternative to OpenTracing for custom instrumentation.

The integration works by hijacking the `OpenTelemetry.getTracerProvider` and `OpenTelemetry.getPropagators` method return values and replacing it with our own.

To enable:
* System Property: `-Ddd.integration.opentelemetry-beta.enabled=true`
* Environment Variable: `DD_INTEGRATION_OPENTELEMETRY_BETA_ENABLED=true`

The actual mapping semantics are subject to change as the API matures.

Events are completely ignored as there's no analogous feature in Datadog APM. This also won't impact the existing MeterProvider.