Config class should be loaded on the bootstrap for instrumentation tests.
`KafkaClientTest` was failing in some cases because `TextMapExtractAdapterTest` didn't extend `AgentTestRunner` and caused `Config` to be loaded on the system classpath before it could be injected into the bootstrap classpath by `KafkaClientTest`.

This adds an assertion in `AgentTestRunner` to help better identify the problem, and makes `TextMapExtractAdapterTest` extend `AgentTestRunner` instead to ensure the classpath is setup correctly regardless of which test executes first.