Final step in introducing stack & handling out-of-order close
Replaced the ArrayList with a ContinuableScope[] in ScopeStack

Dropped the toRestore & depth fields from ContinuableScope
Restoration is now managed via ContinuableScope[]
depth is no longer tracked on the ContinuableScope, but instead tracked by the
topPos member of ScopeStack

In this change, I've taken care to maintain the existing reference counting &
notification semantics -- despite my general feeling that they are not
completely intuitive or correct.

I've also taken care to do minimum of moving code around -- including in tests
to make the change easier to review for correctness.

As such, most things have a 1-to-1 replacement...
- ContinuableScope.depth -> ContinuableScopeManager.ScopeStack.depth
- ContinuableScope.toRestore linked list -> ScopeManager.stack
- ContinuableScopeManager.tlsScope.remove -> ContinuableScopeManager.ScopeStack.clear

The functional part of this change is simply improving out-of-order close behavior.

Previously, given a stack C, B, A...
If B was closed out-of-order, then C would be left on-top.
Once C was closed, B would be restored to the top of stack.
But B would need to close again to proper close A.

If B was not closed twice, then the stack could grow indefinitely.

The new code behaves a bit better -- given a stack C, B, A...
If B is closed out-of-order, then C would left on-top
Once C is closed, the stack will skip over B, since it is already closed -- returning to A.
Then A can be closed properly with an extra close of B.

This doesn't eliminate the possibility of having the stack grow indefinitely,
but I believe it does reduce it significantly.

In hindsight, this could have been achieved via the toRestore field already on
ContinuableScope, but I've elected to implemented a conventional stack instead.

The stack clean-up works by cleaning up the stack on querying active / activeScope.
At access time, the ContinuableScopeManager checks if the ContinuableScope is still active.
If it isn't, then the stack top is moved back a scope at a time until an active
scope is found.

To avoid doing the clean-up often, the stack is proactively unwound on a normal close in
ContinuableScope.close.