POC for collecting finer grained details under a span
The idea being to have a way to collect a much greater amount of data on what is happening without the cost associated with spans.

The first attempt selects "random" classes based on the hashcode of their class name.  This will allow consistency across runs, but is ultimately just a demonstration.  It will result in many methods being instrumented that will never result in useful data being reported to a span.  As such it will incur the cost of a ThreadLocal lookup on potentially hot methods. Future POC's will focus more on identifying better methods.