[PROF-1988] default stackdepth (#1928)
## Change Overview
* sets JFR profiles to use a 256 stack depth by default
* respects user-supplied over-ride for max stack depth
* limits clients to a 1024 max stack depth
* jboss (wildfly) integration tests

## Observations
Based on benchmark results there is little overhead increasing the stackdepth and not truncating typical java web application stacks, which can be regularly greater than 128. 

Due to quirks in JFR initialization (one-shot per JVM, sampling thread started as a side-effect of setting the sampling interval) and issues initializing JMX in some environments, there is a small window in which the agent can set JFR.configure stackdepth

* before profiling starts
* after JMX is initialized

As a result, it's possible the agent may fail to set stackdepth appropriately in some environments. That occurrence is logged for diagnostic purposes. The product will still function as before but some clients may get a warning in the analysis tab (as they do today) that their stacks were truncated with mediation steps.