Allow immediate writes if no remaining pending references
Update the tests... was more complicated to test PendingTraceBuffer since it now requires continuations for root to be delayed.

I removed the test that verifies the trace discarded if span GC'd because it is unreliable... the trace gets buffered and success depends on if GC cleaning happens before the buffer timeout expires.