Make sure DelayQueue internal types are preloaded before starting the tracer's scheduled worker thread.
This avoids a potential loop-back where the worker calling 'take' could trigger a transformation event
which would cause ByteBuddy to do instrumentation matching. That could go on to trigger class loading
in Datadog's class loaders, which may lead to internal exceptions, which are picked up by the exception
profiler.

If this happens when another thread is also throwing an exception that's being profiled the scheduled
worker thread can end up waiting on the exception profiler, while the exception profiler is waiting to
send work to the scheduled worker.