Add missing equals and hashCode method into datadog.trace.instrumentation.opentracing32.OTSpan
It's not possible to check equality of opentracing32.OTSpan intances because it doesn't implement the equals and hashCode methods for the internal agent span. Also the opentracing32.OTSpan class is not part of public API and the getDelegate method is not assessible via the Span or MutableSpan inteface.

At the same time datadog.opentracing.OTSpan has proper implementations of the equals and hashCode methods.