Only consider external dependencies when deciding transitive dependencies to list in the 'dd-trace-ot' pom.
This avoids an issue where a project dependency is listed as a transitive dependency. In some situations, such
as when running a task from a sub-directory instead of the root project, this transitive dependency may not be
resolved yet and has a name of 'unspecified'. Adding it to the 'dd-trace-ot' model leads to a gradle error:

* What went wrong:
Execution failed for task ':compileJava'.
> Could not resolve all files for configuration ':compileClasspath'.
   > Could not find :unspecified:.
     Required by:
         project : > com.datadoghq:dd-trace-ot:0.69.0-SNAPSHOT

Since we're bundling project dependencies we can safely exclude them from the dependencies to add to the pom.