Handle unexpected end of stream explicitly in ProfileUploader
Customers have reported that they see warnings for profile upload
failures in their application logs, coming from `dd-trace-java`.
However, the data is uploaded successfully and is available in the
Datadog app.

We see a similar behaviour internally with services configured to
send profiling data to multiple targets, we see failure warnings
in the logs, but the data is uploaded successfully and available
in the UI.

The cause of the problem is a timeout in the datadog-agent http server
- whenever the profile upload request takes longer than the 5s timeout
we have configured at present, the http server in `datadog-agent`
gracefully closes the connection, `dd-trace-java` sees an empty reply
from server and triggers the UnexpectedEndOfStream exception which is
logged.

This is more likely to happen for customers that have large profiles or
slow network connections to our intake.

This PR changes the logging in ProfileUploader to output a less alarming
message when there is no reply from the server.