Merge NewTaskFor with WrapRunnableAsNewTaskInstrumentation by using a placeholder 'newTaskFor' call. The placeholder call in the advice will be rewritten to call the real 'newTaskFor' at build time.
Note: once the 'NewTaskFor' advice has been rewritten it can only be applied to executors that extend
AbstractExecutorService otherwise we get class verification errors at load time. For other executors
we apply a simpler wrapper that avoids 'newTaskFor'. This results in a bit of duplication but that's
tolerated as it means we can avoid having to inject helpers onto the bootclasspath.