Support instrumentation of classes compiled with a target of 'jsr14'
The 'jsr14' target was a short-term option added between Java 1.4 and Java 6 to help JDK developers test and
migrate code to the generics language feature. It wasn't intended to be used in third-party applications but
it did get used in some OSGi frameworks who wanted to make generic-ready APIs while also supporting Java 1.4

This results in a classfile that has a bytecode version of 48 (Java 1.4) but also contains generic signatures.
Unfortunately this trips up Byte-Buddy when instrumenting because it thinks Java 5+ advice has been incorrectly
rdded to the Java 1.4 class (due to the generic signatures which actually come from the original bytecode).

To fix this we extend Byte-Buddy's TypeConstantAdjustment to also null out generic signatures from bytecode
compiled with a target of 1.4 or below. These signatures should already be null in the original class file,
it's only if the original source used generics and the class was compiled with a target of 'jsr14' that they
would need nullifying.

We also add a config option to switch back to the original Byte-Buddy TypeConstantAdjustment if necessary:

  -Ddd.jsr14.target.adjustment.enabled=false

  DD_JSR14_TARGET_ADJUSTMENT_ENABLED=false