Servlet Async Dispatch should not result in another servlet.request span
`servlet.request` spans are often given special meaning and should be 1:1 for the number of actual requests, not based on internal routing of a request.

The new servlet.dispatch span is activated where the dispatched request would have been created, so the duration encompasses the request duration.

A minor "breaking" change:
* the `servlet.dispatch` tag is no longer added to the `servlet.request` span.