Tomcat 5.5+ instrumentation
This instrumentation overrides the generic servlet instrumentation and provides a more holistic view of the actual lifecycle of a tomcat request. Particularly important is the better detection of final http status code when an exception is thrown and a different status is returned via an exception handler.

Request/Response objects are always recycled so we use that as the last possible point to ensure the most accurate status code.

(The release notes should indicate that this is a pretty significant change in instrumentation for anyone using Tomcat and to be aware when upgrading. This includes possible impact to duration metrics due to an increased measurement scope.)