Avoid potential ClassCastException in AWSv1 integration
The different AWS client libraries (such as core and s3) may be deployed in separate classloaders
on modular application servers like JBoss/Wildfly. This can result in the AWSv1 instrumentation
applying to different classloaders. For example AWSHttpClientInstrumentation will only apply to
the core library while RequestInstrumentation could apply to the s3 library (among others) as its
matcher covers any class that extends AmazonWebServiceRequest.

Both instrumentations previously asked for RequestMeta to be injected as a helper so it could be
used as the value-type in the shared context-store. On modular appservers that enforce visibility
this could result in RequestMeta being defined twice, once in each classloader. This could later
result in a ClassCastException when the tracing handler attempts to retrieve RequestMeta added by
RequestInstrumentation, because the two RequestMeta types are different even though they have the
same name.

To avoid this potential issue we use a simple Map instead of the RequestMeta holder to store the
request metadata in the context store. We use the "aws...." tag names as the keys for consistency.

We also decouple TracingRequestHandler from AwsSdkClientDecorator so the handler can be used in
AWSHttpClientInstrumentation without needing to drag in all the AwsSdkClientDecorator types.