Enable test to bind to available open port
Instead of picking an available port which may get used
between the call to find it and the call to use it use
port 0 so the system can pick an available port. Fetch
the port used out of the Listener.

Remove the PassThroughHttpListener subclass since we
no longer require it to choose a random port.