Log PendingTrace state after span finish
When troubleshooting async tracing behavior, it is helpful to understand the state of the pending trace.  Added that extra bit of info to the span finished log message rather than create a new log.

I also added a debug message to report that a trace is being discarded when using `strictTraceWrites`.