Rework how JFR events are disabled based on JVM version (#2858)
* Include all bin/ folder to .gitignore

* Rework how JFR events are disabled based on JVM version

We want to disable events enabled by default in dd.jfp based on the
version. For example, OldObjectSample or NativeMethodSample have poor
performance on older versions. We don't want to penalize customers who
are on these older versions and just use the default profile.

However, we still want to allow users to override any settings they
want as they may have perfectly valid reasons to do so and to accept to
pay the cost for them.