Add printing mode for TimelineCheckpointer
This mode is used for generating test cases for the profiling backend.

The generated files are parsed for test cases, and the profiling backend
uses them to verify that it can reconstruct the timeline appropriately.
We need "real world" scenarios for the profiling backend to support
every possible cases of valid (or not so valid) sequence of checkpoints.