Remove the SuspendResumeValidator
A balance of suspend/resume can't be enforced because the following
sequences are valid:
- resume | suspend
- resume | suspend | endTask
- resume | endTask <- that's the problematic case

In this last case, because we don't have a suspend after the resume, the
balance is broken. However, this is an acceptable case as both suspend
and endTask are `END` events and signal the end of the execution of this
span on this thread.