Dump information when a profile fails to upload because it's too big (#3046)
* Merge ResponseCallback into Callback

* Dump information when a profile fails to upload because it's too big

When receiving a 413 error code, we only know that the profile is too
big to be accepted by the backend. However, we have no idea why the
profile is too big.

With this change, we try to extract information about the size of the
events and log it to the command line. This allows us to ask customers
for more informations, and react accordingly. For example, if the
profile is too big due to too many exceptions being thrown and
overloading the profile, we can advise the customer to disable the
Exception profiler.

* Fixes according to review

* Fix check

* Fix spotless

* Make jacoco happy

* Delete tmp file when done with it

* Add config to enable upload summary