Setting up codescanning for dd-trace-java
Static code analysis on dd-trace-java using Github's Code Scanning.

This PR has Github workflow configs to trigger the Code Scanning on every PR and on push to default branch. As of now only default security based CodeQL queries are configured. We can gradually add queries for Code Quality later. The SARIF results for the code-scanning job is uploaded as an artifact for reference.

The alerts will not show up on master branch until the PR gets merged.