Implement 'closePrevious' and 'activateNext' to help track potentially open-ended iterations (#3194)
When iterating over a potentially open-ended iteration, such as polling messages from a queue:

* Call `AgentTracer.closePrevious(finishSpan)` just before receiving the message
    set `finishSpan` to false if you intend to finish spans on a different condition, such as on a commit
* Create a new span for consuming the message
* Call `AgentTracer.activateNext(messageSpan)` to activate that span and register it for cleanup

There's no need to store the resulting scope - it will be automatically closed either when the next message is received, when a surrounding non-iterating scope is closed, or when the span times out. A background task is only scheduled when there's no surrounding non-iterating scope to detect when the iteration scope should be closed, or when a custom scope manager is used.

The "keep-alive" time for iteration spans can be adjusted with this system property:

-Ddd.trace.scope.iteration.keep.alive=<seconds>

or by setting this environment variable

DD_TRACE_SCOPE_ITERATION_KEEP_ALIVE=<seconds>

Setting the timeout to zero disables the background task.