Support custom scope managers with OT33
To do this we avoid deprecated/removed methods as much as possible even if it means taking
a more circuitous route. We also need to support having an active span but no active scope
when querying the scope manager because the API to retrieve the scope was removed in OT33.