Expand RejectedExecutionTest to cover rejected calls to 'execute'
The tracer wraps runnables passed into 'execute' with its own wrapper to track propagation,
but we should remove this wrapper when the task is rejected so the RejectedExecutionHandler
sees the original runnable (some frameworks rely on this behaviour.)

Note that runnables passed to 'submit' do not use our wrapper because they are wrapped with
FutureTask instead. The expected behaviour there is that the FutureTask wrapper is passed to
the RejectedExecutionHandler.