Util to get stacktrace for reported vulnerabilities
What Does This Do
Util to do partial walks of stacktraces. Excludes elements that belong to the dd-trace-java project.

Motivation
In IAST, when we detect a vulnerability, we need to get the class and line where it happened. The relevant class and the line will not be the last in the stack. The relevant class and line will be calculated ignoring internal dd-trace-java classes, usually applying a filter of classes on top of that (e.g standard library).
So the filtering of dd-trace-java elements is built-in, and the API uses streams for further filtering operations.
In most cases, only a few elements need to be walked through (e.g. 2-4), and only in some rare cases the whole stack needs to be examined.

Additional Notes
This functionality will be used whenever a vulnerability is found, so its performance is critical to vulnerability reporting. To optimize performance, three implementations are included:

Default implementation: uses Thread.currentThread().getStackTrace() and gets an array with the full stacktrace. This is the most expensive when doing a partial walk, since it needs to create StackTraceElement objects for the whole stacktrace.
Improve performance for JDK8 with HotSpot ((JavaLangAccess)). Relies on JavaLangAccess to create StackTraceElements on-demand.
Improve performance for JDK9+ (StackWalker). Uses the newer StackWalker API.
The implementation is currently in the internal-api-* packages. This was picked arbitrarily to some extent, so other suggestions about locations are welcome. Initially, this API will only be used from the upcoming dd-java-agent/iast module

Future Work
Future versions of this utility may include further performance optimizations:
 - Filtering by class name before creating StackTraceElement objects (possible in all implementations except the default).
 - Skipping a fixed number of initial frames without inspecting them at all. We will always have a number of elements that are part of dd-trace java, and we can skip them by default. This would save some superfluous allocations of StackTraceElements.