IAST Overhead control engine V0
Implement Overhead Control Engine (OCE) for IAST operations. It creates
a module that by design ensures the maximum overhead for IAST
operations. The OCE will be responsible for assuring performance and
short circuiting the logic if it reaches the maximum.

Low overhead is a mandatory requirement for a software that is running
in production. The main constraint is that having to introduce many
hooks for taint tracking keeping a low overhead is a challenging task.
The OCE is an element that by design ensures that the overhead does not
go over a maximum limit. It will measure operations being executed in a
request and it will deactivate detection (and therefore reduce the
overhead to nearly 0) if a certain threshold is reached.

First approach sets threshold to:
- Only about 30% of the requests will be analysed
- 2 concurrent request
- 2 vulnerabilities per request
- 2 vulnerabilities out of request

Future versions will implement a more fine-tuned policy for different
operation types. The current PR implements one of the simplest policies.

Co-authored-by: Santiago Mola <santiago.mola@datadoghq.com>