Fix issue in default stack walker implementation
`Thread.currentThread().getStackTrace()` seems to be the same as `new
Throwable.getStackTrace()` for the general case, but it results in an
additional frame. This might result in a slight performance loss, and
also was the source of a bug where we got an unexpected result.

Refactored stack walker tests to spot this kind of issue.