Fix debugger smoke test flakiness
TracerDebuggerIntegrationTest seems to face a race between Tracer
instrumentation and Debugger instrumentation:
while a class is loading
(org.springframework.web.servlet.DispatcherServlet), Tracer is
transforming it while Debugger receiving configuration and add new
transformer, but this transformer will not be take into account
because class is already transforming and after that will be
considered as loaded.
Debugger perform also a retransformation for already loaded classes
but as this class is currently transforming with Tracer it is not part
of loaded classes.
Workaround is to try to re-instrument the class.

While topping Server app for ServerDebuggerIntegrationTest, Http
server could be stopped before sending Http response. Catch exception
as it is not an issue.