[IAST] OverheadController to add caps on analyzed requests and reported vulnerabilities (#3694)
The OverheadController ensures that the overhead does not
go over a maximum limit. It will measure operations being executed in a
request and it will deactivate detection (and therefore reduce the
overhead to nearly 0) if a certain threshold is reached.

First approach sets threshold to:
- Only about 30% of the requests will be analysed
- 2 concurrent request
- 2 vulnerabilities per request
- 2 vulnerabilities out of request

Future versions will implement a more fine-tuned policy for different
operation types. The current PR implements one of the simplest policies.

Co-authored-by: Santiago Mola <santiago.mola@datadoghq.com>