strip 'v' from version tag
Currently the user must remove the 'v' from Github's suggested default release title like "v0.123.0" -> "0.123.0" and put this new value in the 'release title' field which populates $VERSION.  If the user does not strip the 'v', subsequent release steps fail.  This change gets the VERSION from the tag instead and strips the leading 'v'.  The user should still enter a name, but if they don't remove the 'v', it should be ok or they can just rename it later.