Implement log probe instrumentation
Reuse snapshot instrumentation but with custom summary builder
allowing to generate a message as log.
variables are resolved reusing ValueScript class.
Right now, logs will generate the same data than regular snapshots,
but we will remove that later to just the data required for filling
the template variable.
Error during evaluation will be also reported  in the snapshot in a
later PR