Fix side effects in CallSiteUtils causing bug
Fixes APPSEC-6856.

The problem occurs when some call site advice for
a) an instance method
b) that is not interested in the receiver of the call (doesn't use `@This`),
c) doesn't use all the arguments (or not all arguments in order), and
d) there is no stack manipulation capable of duplicating the required
arguments in the stack without using an intermediary array.

In this case, the following code is executed:

```
public static void dup(MethodVisitor mv, Type[] argumentTypes, int[] indices) {
  int[] opcodes = lookupParamDupOpcodes(argumentTypes, indices);
  if (opcodes == null) {
    pushArray(mv, argumentTypes.length, argumentTypes);
    loadArray(mv, argumentTypes.length, argumentTypes);
    loadArray(mv, argumentTypes, indices);
    mv.visitInsn(POP); // pop out array
  } else {
    // ...
  }
```

The problem is that `lookupParamDupOpcodes` modifies the argument
`indices`, and therefore its content is wrong on the last call to
`loadArray()`.