Fix inner classes matching for line probes
When class are already loaded, we need to retransform them for passing
through the DebuggerTransformer.
But to detect which class need to be retransformed, we rely on the
naming. However inner classes or Top-Level classes are in independent
class files and the only way to associate them is the sourcefile
attribute in each class file.
So we need to track all dependent classes by their source file in
order to avoid scanning all loaded classes to parse their sourcefile
and try to match with actual probe definitions.