Fix incorrect arg name for already traced methods
When instrumenting a method with a debugger probe that contains also
tracer instrumentation (using @trace annotation or
`trace.methods` property) argument names was not mapped correctly and
p0, p1, appears in the snapshot.
This is due to ByteBuddy transformation messing up the
LocalVariableTable that helps to maps local vars with names. Args
are also local vars and following the JVM spec, the first local vars
must match the methods args.
LocalVariableTable is optional and only used for debugger so it does
not prevent the actual execution of the method.
The fix consists to reassigned local var information to arg slot.
Added smoke tests for this case.