Simplify scope activation in AWS-SDK instrumentations.
Avoid storing the scope in the v1 attributes; instead close any active AWS-SDK span when the client
finishes its execution, like we do for v2. Also introduce a no-op scope when calling out from AWS to
Netty in async mode to make sure that it can't accidentally capture continuations which incorrectly
keep the AWS trace alive. (We close the AWS scope but there may be other scopes left on the stack.)