Refactor data collection instrumentation
Main goal of this refactoring is to support duplicated probe as first
class citizen instead of having it on the side lines.
We shift the way snapshots are constructed to focus on CapturedContext
as main data collection object and evaluation of log template and
condition. The commit phase is considering the captured contexts and
decide if we need to send a snapshot and build it by probe with all
information required (entry/exit/line context, duration, evaluation
errors).
Each context has a status compiling information about the evaluation
results and help make the decision at commit time. Those statuses are
per probe.