Refactor context evaluation & ProbeDetails
Delegate context evaluation depending on the probe type. So we
introduced an evaluate method in ProbeDefinition and move current code
into LogProbe. Other kinds of probe can then evaluate the context
differently depending on what would need to be evaluated.
Refactor ProbeDetails to be just an interface so ProbeDefinition
will implement is instead of creating a custom class containing all
information that are required for the context evaluation.
Remove ProbeDefinition.MethodLocation to keep only one in bootstrap
project (Used by snapshot ans CapturedContext.