Avoid duplicate elements in TaintedMap
In TaintedMap, insert at the end of the chain, instead of at the head.
This allows us to test for duplicates. This prevents the pathological
case where the same object is tainted many times, making the linked list
of a bucket have an unbound length and distorting the size estimate.

This should be, in theory, slower. But microbenchmarks do not show any
noticeable difference. Perhaps because given our parameters, less than
0.2% of the chains are expect to be longer than 2 elements.