Adapt spark instrumentation for databricks (#5118)
Databricks is using Spark in a slightly different way than other cluster managers, with a model of one cluster = one spark application, whereas one cluster can have multiple Spark applications on YARN or Kubernetes.

The spark driver, and consequently the Spark application, are long running and bound to the cluster with spark jobs (not applications) submitted to it. In this case, we make spark jobs the local root spans.