[Spark] Instrument YARN application end to not drop traces (#5267)
Instrument the spark yarn.ApplicationMaster.finish() function in order to
- catch the exit code and exit message of the application
- make sure the spark.application span is finished before the trace processor is exited

## Capturing exit code

An application can fail during spark computations (when performing operations on the data) or in the user code (like trying to select a column that does not exists)

Failures during spark computations are caught by the SparkListener, but not the errors in the user code so instrumenting the finish function allows to also capture those

## Finish the spark.application span

The spark.application span is the local root span and is finished when the event onApplicationEnd is received. However if the sparkContext is not stopped explicitly by the user (using spark.stop()), spark is stopped in a shutdown hook. In this case, the onApplicationEnd event can be received after the trace processor was exited, leading to the trace being dropped, with the following logs

```
23/05/17 13:11:23 INFO ApplicationMaster: Final app status: SUCCEEDED, exitCode: 0
23/05/17 13:11:23 INFO SparkContext: Invoking stop() from shutdown hook
[dd.trace 2023-05-17 13:11:23:205 +0000] [dd-trace-processor] DEBUG datadog.trace.agent.common.writer.TraceProcessingWorker - Datadog trace processor exited. Publishing traces stopped
[dd.trace 2023-05-17 13:11:23:212 +0000] [spark-listener-group-shared] DEBUG datadog.trace.agent.common.writer.RemoteWriter - Trace written after shutdown.. Counted but dropping trace
```

By manually calling the function behind onApplicationEnd, we can make sure that the trace is always finished before the tracer is exited